Fixed ceres to not run when on windows
