+++ src/sonar_tracing/include/sonar_tracing/snell_ray_tracing.h
@@ -13,31 +13,7 @@
 #define SNELL_RAY_TRACING_H
 
 #include <Eigen/Dense>
-#include <ceres/ceres.h>
 
-class LayerWidthCostFunctor {
-public:
-    LayerWidthCostFunctor(double height, double speed) : height(height), speed(speed)
-    {
-    }
-
-    template <typename T>
-    bool operator()(const T* const x1, const T* const x2, T* e) const
-    {
-        e[0] = pow((*x2-*x1)*(*x2-*x1)+T(height*height), 0.25)/T(sqrt(speed));
-        return true;
-    }
-    
-    static ceres::CostFunction* Create(double height, double speed)
-    {
-        return new ceres::AutoDiffCostFunction<LayerWidthCostFunctor, 1, 1, 1>(
-            new LayerWidthCostFunctor(height, speed));
-    }
-
-private:
-    double height;
-    double speed;
-};
 
 // here, we assume that the origin is at (0, 0), and the points all have x > 0
 std::pair<Eigen::VectorXd, Eigen::MatrixXd> trace_multiple_layers(const Eigen::VectorXd& layer_depths, const Eigen::VectorXd& layer_speeds, const Eigen::MatrixXd& end_points);
