+++ src/sonar_tracing/src/snell_ray_tracing.cpp
@@ -15,6 +15,10 @@
 #include <opencv2/highgui/highgui.hpp>
 #include <opencv2/imgproc/imgproc.hpp>
 
+#ifndef _MSC_VER
+  #include <ceres/ceres.h>
+#endif
+
 using namespace std;
 
 // here, we assume that the origin is at (0, 0), and the points all have x > 0
@@ -35,6 +39,30 @@ pair<Eigen::VectorXd, Eigen::MatrixXd> trace_multiple_layers(const Eigen::Vector
 
 pair<double, Eigen::VectorXd> trace_single_layers(const Eigen::VectorXd& layer_depths, const Eigen::VectorXd& layer_speeds, const Eigen::Vector2d& end_point)
 {
+#ifndef _MSC_VER
+    class LayerWidthCostFunctor {
+    public:
+        LayerWidthCostFunctor(double height, double speed) : height(height), speed(speed)
+        {
+        }
+
+        template <typename T>
+        bool operator()(const T* const x1, const T* const x2, T* e) const
+        {
+            e[0] = pow((*x2-*x1)*(*x2-*x1)+T(height*height), 0.25)/T(sqrt(speed));
+            return true;
+        }
+    
+        static ceres::CostFunction* Create(double height, double speed)
+        {
+            return new ceres::AutoDiffCostFunction<LayerWidthCostFunctor, 1, 1, 1>(
+                new LayerWidthCostFunctor(height, speed));
+        }
+
+    private:
+        double height;
+        double speed;
+    };
     Eigen::VectorXd layer_widths(layer_depths.rows()+2);
     layer_widths(0) = 0.;
 
@@ -92,6 +120,8 @@ pair<double, Eigen::VectorXd> trace_single_layers(const Eigen::VectorXd& layer_d
 
     // return the final ray time and the optimized intermediary points
     return make_pair(2.*summary.final_cost, layer_widths);
+#endif
+    return make_pair(0., Eigen::VectorXd::Zero(layer_depths.rows()));
 }
 
 void visualize_rays(const Eigen::MatrixXd& end_points, const Eigen::VectorXd& layer_depths,
